refactor: Migrate orbit-agent tests to per-module tests/ layout [ORB-00220]
