+++ crates/orbit-agent/src/loop_engine/tool_dispatch/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod schema;
