+++ crates/orbit-agent/src/providers/grok/grok_cli.rs
@@ -32,13 +32,4 @@ impl GrokCliTransport {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn grok_args_pass_model_with_long_flag() {
-        let transport = GrokCliTransport::new(Some("grok-build".to_string()));
-
-        assert_eq!(transport.args(), vec!["--model", "grok-build"]);
-    }
-}
+mod tests;
