+++ crates/orbit-agent/src/providers/grok/grok_cli/tests/args.rs
@@ -1,0 +1,10 @@
+#![allow(missing_docs)]
+
+use super::super::*;
+
+#[test]
+fn grok_args_pass_model_with_long_flag() {
+    let transport = GrokCliTransport::new(Some("grok-build".to_string()));
+
+    assert_eq!(transport.args(), vec!["--model", "grok-build"]);
+}
