+++ crates/orbit-agent/src/providers/grok/grok_cli/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod args;
