+++ crates/orbit-agent/src/types/response/envelope/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod parse;
