+++ crates/orbit-agent/src/types/response/usage/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod sum;
